Add files via upload
I made the main menu with the options in a loop, so the user can return to it whenever they want. I also added the settings option, where the user can change the number of the given letters, as well as the number of rounds. I added and the option of adding words to the dictionary. I only need to build this part of the program which checks whether the made-up word, which consists of repeating letters is a valid input based on the given letters and to check it in the dictionary and add to the points.