Add files via upload
This is the last edit, which took several days to be meaningful and complete. 
First of all, I eliminated the check, whether the input word was longer than the actual available letters. Instead, I tweaked the check for repetition of letters, which I have had in mind. I added some variable which with the needed value activate a certain mechanism which loops the whole code until a proper input is given. After that, I changed the whole part of the code that checked whether the input word actually exists. I changed it, so it got the job done. Then I added a score counter as well. And the code was pretty much ready. Final tweaks and I added some pretty descriptive comments, so everything is clear.